+++ projects/net.splitcells.dem/src/main/java/net/splitcells/dem/utils/FodsUtility.java
@@ -16,11 +16,13 @@
 package net.splitcells.dem.utils;
 
 import net.splitcells.dem.lang.annotations.JavaLegacyArtifact;
+import net.splitcells.dem.lang.perspective.Perspective;
 import org.w3c.dom.Element;
 
 import static net.splitcells.dem.lang.Xml.*;
 import static net.splitcells.dem.lang.namespace.NameSpaces.FODS_TABLE;
 import static net.splitcells.dem.lang.namespace.NameSpaces.FODS_TEXT;
+import static net.splitcells.dem.lang.perspective.PerspectiveI.perspective;
 import static net.splitcells.dem.utils.ConstructorIllegal.constructorIllegal;
 
 /**
@@ -39,5 +41,10 @@ public class FodsUtility {
         tableCell.appendChild(cellContentElement);
         cellContentElement.appendChild(textNode(cellContent));
         return tableCell;
+    }
+
+    public static Perspective tableCell2(String cellContent) {
+        return perspective("table-cell", FODS_TABLE)
+                .withChild(perspective("p", FODS_TEXT).withChild(perspective(cellContent)));
     }
 }